chore: move items from storage to install and small cleanup
