+++ install/package.json
@@ -3,6 +3,7 @@
   "description": "MO CMS Demo",
   "version": "0.0.1",
   "scripts": {
+    "cms": "php composer.phar update && cd vendor/maxtream/themages",
     "build": "cd vendor/maxtream/themages && yarn install && gulp build",
     "watch": "cd vendor/maxtream/themages && gulp watch"
   },
